Ft/dh 5037 kula project (#321)
* Fb/dh 5044 (#314)

* fb/DH-5044_first commit add overall key module with add route to generate a new key

* fb/DH-5044 add one way hashing logic to API keys

* fb/DH-5044 add key validation to GET instruction route

* fb/DH-5044 add salt to enf file

* fb/DH-5044 fix linting issues

* DH-5041 refactor for project Kula

* DH-5041 migration script for project Kula

* DH-5041 added datetime field and pagination for resources

* refactor front-end to align with refactored enterprise

* fixed admin console interfaces

* fix admin console and slack

* added validation for generation endpoints

* admin-console prettier fixed

* updated dataherald submodule

* DH-5041 fixed migration script

* fix build problem about admin user update to work with full user instead of partial updates with patch

* DH-5037 fix aggregated generation and fix migration script

* fix field renaming in the backend

* DH-5037 fix golden sql tests

* DH-5037 switch dataherald submodule to kula branch

* fix confidence score

* [ai][server] Create an initial script to generate data (#322)

* [ai][server] Create an initial script to generate data

* Add makefile

* Revert "fix confidence score"

This reverts commit 03b0045aae2890375f18211a3e4597cff7a47951.

* DH-5037 fixed migration script

---------

+++ apps/ai/server/modules/table_description/service.py
@@ -1,7 +1,9 @@
 import httpx
+from fastapi import HTTPException, status
 
 from config import settings
 from modules.db_connection.service import DBConnectionService
+from modules.organization.service import OrganizationService
 from modules.table_description.models.requests import (
     ScanRequest,
     TableDescriptionRequest,
@@ -11,92 +13,86 @@ from modules.table_description.models.responses import (
     DatabaseDescriptionResponse,
     TableDescriptionResponse,
 )
+from modules.table_description.repository import TableDescriptionRepository
 from utils.exception import raise_for_status
+from utils.misc import reserved_key_in_metadata
 
 
 class TableDescriptionService:
     def __init__(self):
+        self.repo = TableDescriptionRepository()
+        self.org_service = OrganizationService()
         self.db_connection_service = DBConnectionService()
 
     async def get_table_descriptions(
-        self, table_name: str, db_connection_id: str
+        self, table_name: str, org_id: str
     ) -> list[TableDescriptionResponse]:
-        async with httpx.AsyncClient() as client:
-            response = await client.get(
-                settings.engine_url + "/table-descriptions",
-                params={"db_connection_id": db_connection_id, "table_name": table_name},
-                timeout=settings.default_engine_timeout,
-            )
-            raise_for_status(response.status_code, response.text)
+        organization = self.org_service.get_organization(org_id)
+        table_descriptions = self.repo.get_table_descriptions(
+            organization.db_connection_id, table_name
+        )
 
-            table_descriptions = [
-                TableDescriptionResponse(**td) for td in response.json()
-            ]
-
-            for table_description in table_descriptions:
-                for column in table_description.columns:
-                    column.categories = (
-                        sorted(column.categories) if column.categories else None
-                    )
-
-            return table_descriptions
-
-    async def get_table_description(
-        self, table_description_id: str
-    ) -> TableDescriptionResponse:
-        async with httpx.AsyncClient() as client:
-            response = await client.get(
-                settings.engine_url + f"/table-descriptions/{table_description_id}",
-                timeout=settings.default_engine_timeout,
-            )
-            raise_for_status(response.status_code, response.text)
-
-            table_description = TableDescriptionResponse(**response.json())
+        for table_description in table_descriptions:
             for column in table_description.columns:
                 column.categories = (
                     sorted(column.categories) if column.categories else None
                 )
+        return table_descriptions
 
-            return table_description
-
-    async def get_database_table_descriptions(self, db_connection_id: str):
-        if not db_connection_id:
-            return []
-
-        async with httpx.AsyncClient() as client:
-            response = await client.get(
-                settings.engine_url + "/table-descriptions",
-                params={"db_connection_id": db_connection_id, "table_name": ""},
-                timeout=settings.default_engine_timeout,
-            )
-            raise_for_status(response.status_code, response.text)
-
-            db_connection = self.db_connection_service.get_db_connection(
-                db_connection_id
+    async def get_table_description(
+        self, table_description_id: str, org_id: str
+    ) -> TableDescriptionResponse:
+        if not self.table_description_in_organization(table_description_id, org_id):
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail="Table Description not found",
             )
 
-            table_descriptions = [
-                TableDescriptionResponse(**td) for td in response.json()
-            ]
-            tables = [
-                BasicTableDescriptionResponse(
-                    id=td.id,
-                    name=td.table_name,
-                    columns=[c.name for c in td.columns],
-                    sync_status=td.status,
-                    last_sync=str(td.last_schema_sync) if td.last_schema_sync else None,
-                )
-                for td in table_descriptions
-            ]
-            return [
-                DatabaseDescriptionResponse(
-                    alias=db_connection.alias,
-                    tables=tables,
-                    db_connection_id=db_connection_id,
-                )
-            ]
+        table_description = self.repo.get_table_description(table_description_id)
 
-    async def sync_table_descriptions_schemas(self, scan_request: ScanRequest) -> bool:
+        for column in table_description.columns:
+            column.categories = sorted(column.categories) if column.categories else None
+
+        return table_description
+
+    async def get_database_table_descriptions(self, org_id: str):
+        organization = self.org_service.get_organization(org_id)
+        db_connection = self.db_connection_service.get_db_connection(
+            organization.db_connection_id, org_id
+        )
+        table_descriptions = self.repo.get_table_descriptions(db_connection.id)
+
+        tables = [
+            BasicTableDescriptionResponse(
+                id=td.id,
+                name=td.table_name,
+                columns=[c.name for c in td.columns],
+                sync_status=td.status,
+                last_sync=str(td.last_schema_sync) if td.last_schema_sync else None,
+            )
+            for td in table_descriptions
+        ]
+        return [
+            DatabaseDescriptionResponse(
+                alias=db_connection.alias,
+                tables=tables,
+                db_connection_id=db_connection.id,
+            )
+        ]
+
+    async def sync_table_descriptions_schemas(
+        self, scan_request: ScanRequest, org_id: str
+    ) -> bool:
+        db_connection = self.db_connection_service.get_db_connection(
+            scan_request.db_connection_id, org_id
+        )
+
+        if not db_connection:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail="Database connection not found",
+            )
+
         async with httpx.AsyncClient() as client:
             response = await client.post(
                 settings.engine_url + "/table-descriptions/sync-schemas",
@@ -110,19 +106,44 @@ class TableDescriptionService:
         self,
         table_description_id: str,
         table_description_request: TableDescriptionRequest,
+        org_id: str,
     ) -> TableDescriptionResponse:
+        reserved_key_in_metadata(table_description_request.metadata)
+        if not self.table_description_in_organization(table_description_id, org_id):
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail="Table Description not found",
+            )
+
         async with httpx.AsyncClient() as client:
-            response = await client.patch(
+            response = await client.put(
                 settings.engine_url + f"/table-descriptions/{table_description_id}",
                 json=table_description_request.dict(exclude_unset=True),
             )
             raise_for_status(response.status_code, response.text)
             return TableDescriptionResponse(**response.json())
 
-    async def delete_table_description(self, table_description_id: str):
+    async def delete_table_description(self, table_description_id: str, org_id: str):
+        if not self.table_description_in_organization(table_description_id, org_id):
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail="Table Description not found",
+            )
+
         async with httpx.AsyncClient() as client:
             response = await client.delete(
                 settings.engine_url + f"/table-descriptions/{table_description_id}",
             )
             raise_for_status(response.status_code, response.text)
             return True
+
+    def table_description_in_organization(
+        self, table_description_id: str, org_id: str
+    ) -> bool:
+        table_description = self.repo.get_table_description(table_description_id)
+
+        db_connection = self.db_connection_service.get_db_connection(
+            table_description.db_connection_id, org_id
+        )
+
+        return True if db_connection else False
